requirements.txt
asgiref==3.4.1
django==3.2.14
flake8==4.0.1
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
gunicorn==20.1.0
pytest-django==4.5.2
python-decouple==3.6
dj-database-url==0.5.0
psycopg2-binary==2.9.3
django-s3-folder-storage==0.5
codecov==2.1.12
pytest-cov==3.0.0
Collectfast==2.2.0
Pillow==9.1.1
model-bakery==1.6.0
six==1.16.0
django-taggit==3.0.0